+++ main-service/src/main/java/ru/practicum/service/events/EventsPrivateServiceImpl.java
@@ -16,9 +16,9 @@ import ru.practicum.dto.requests.RequestStatus;
 import ru.practicum.exception.EntityForbiddenException;
 import ru.practicum.exception.EntityNotFoundException;
 import ru.practicum.mapper.CategoryMapper;
-import ru.practicum.mapper.EventFullMapper;
-import ru.practicum.mapper.EventNewMapper;
-import ru.practicum.mapper.EventShortMapper;
+import ru.practicum.mapper.event.EventFullMapper;
+import ru.practicum.mapper.event.EventNewMapper;
+import ru.practicum.mapper.event.EventShortMapper;
 import ru.practicum.mapper.LocationMapper;
 import ru.practicum.mapper.RequestMapper;
 import ru.practicum.model.Event;
